core: fix stateless VM which was reverted to ram snapshot
We were removing the memory volumes from snapshots in case the live part
of the snapshot did not happen and the snapshot is set to include
memory. This logic is not good because when it comes to stateless
snapshots, the snapshots contain memory (the memory that was associated
with the active snapshot) and it is not a live snapshot, but we don't
want to remove the memory associated with it.

Therefore, this patch changes the logic that is used to determine
whether or not to remove the memory which is associated with snapshot
when the live part of the snapshot doesn't happen, such that we'll also
ask if the memory was created as part of this snapshot and only if it
is, we will delete it.

Bug-Url: https://bugzilla.redhat.com/1004184

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/CreateAllSnapshotsFromVmCommand.java
@@ -229,9 +229,8 @@ public class CreateAllSnapshotsFromVmCommand<T extends CreateAllSnapshotsFromVmP
             } else {
                 // If the created snapshot contains memory, remove the memory volumes as
                 // they are not going to be in use since no live snapshot is created
-                if (createdSnapshot.containsMemory()) {
+                if (getParameters().isSaveMemory() && createdSnapshot.containsMemory()) {
                     logMemorySavingFailed();
-
                     getSnapshotDao().removeMemoryFromSnapshot(createdSnapshot.getId());
                     removeMemoryVolumesOfSnapshot(createdSnapshot);
                 }
@@ -241,7 +240,9 @@ public class CreateAllSnapshotsFromVmCommand<T extends CreateAllSnapshotsFromVmP
                 revertToActiveSnapshot(createdSnapshot.getId());
                 // If the removed snapshot contained memory, remove the memory volumes
                 // Note that the memory volumes might not have been created
-                removeMemoryVolumesOfSnapshot(createdSnapshot);
+                if (getParameters().isSaveMemory() && createdSnapshot.containsMemory()) {
+                    removeMemoryVolumesOfSnapshot(createdSnapshot);
+                }
             } else {
                 log.warnFormat("No snapshot was created for VM {0} which is in LOCKED status", getVmId());
             }
